pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "matplotlib>=3.10.3",
    "mplfinance>=0.12.10b0",
    "pandas>=2.3.1",
    "patterns>=0.1.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ta>=0.11.0",
    "telegram>=0.0.1",
]
